Converted volume levels to Mediaportal default (0-100 -> 0-65535). Cleaned up and reverted a few unnecessary changes.

+++ mediaportal/MiniDisplayLibrary/MiniDisplayPlugin/Drivers/DM140GINK.cs
@@ -151,7 +151,7 @@ namespace MediaPortal.ProcessPlugins.MiniDisplayPlugin.Drivers
         vfd.writeLine(line, RemoveDiacritics(message));
 
         //Volume calculation
-        int vol = (int)VolumeHandler.Instance.Volume / 2730;
+        int vol = VolumeHandler.Instance.Volume / 2730;
 
         // Display symbols
         // LiveTV, LiveRadio
